+++ src/dplx/dlog/bus/buffer_bus.cpp
@@ -1,0 +1,12 @@
+
+// Copyright Henrik S. Gaßmann 2021-2023.
+//
+// Distributed under the Boost Software License, Version 1.0.
+//         (See accompanying file LICENSE or copy at
+//           https://www.boost.org/LICENSE_1_0.txt)
+
+#include "dplx/dlog/bus/buffer_bus.hpp"
+
+namespace dplx::dlog
+{
+}
